MCR-2178 fix JWT claim mcr:ip for IPv6
moved code to detect session steal from MCRServlet to MCRFrontendUtil, so
it can be used by restapi, too.
